Move help and uses to their own folders
