+++ src/routes/uses/index.json.ts
@@ -1,5 +1,5 @@
 import type { Request, Response } from 'express';
-import { parseFile } from '../utils/parser';
+import { parseFile } from '../../utils/parser';
 
 export function get(_: Request, res: Response) {
 	const article = parseFile('content/uses.md', (data: any, content: string) => {
